+++ conda_forge_tick/auto_tick.py
@@ -44,7 +44,7 @@ from urllib.error import URLError
 import github3
 from uuid import uuid4
 
-from conda_forge_tick.utils import pushd
+from conda_forge_tick.os_utils import pushd
 from conda_forge_tick.lazy_json_backends import (
     get_all_keys_for_hashmap,
     LazyJson,
@@ -61,6 +61,7 @@ from conda_forge_tick.git_utils import (
     is_github_api_limit_reached,
     comment_on_pr,
 )
+from conda_forge_tick.os_utils import eval_cmd
 from conda_forge_tick.utils import (
     setup_logger,
     pluck,
@@ -68,7 +69,6 @@ from conda_forge_tick.utils import (
     dump_graph,
     CB_CONFIG,
     parse_meta_yaml,
-    eval_cmd,
     sanitize_string,
     frozen_to_json_friendly,
     yaml_safe_load,
