+++ conda_forge_tick/migrators/cos7.py
@@ -4,7 +4,7 @@ import typing
 from typing import Any
 from ruamel.yaml import YAML
 
-from conda_forge_tick.utils import pushd
+from conda_forge_tick.os_utils import pushd
 from conda_forge_tick.migrators.core import MiniMigrator
 
 if typing.TYPE_CHECKING:
