+++ conda_forge_tick/migrators/license.py
@@ -8,8 +8,8 @@ import logging
 
 from rever.tools import replace_in_file
 
-from conda_forge_tick.utils import pushd
-from conda_forge_tick.utils import eval_cmd, _get_source_code
+from conda_forge_tick.os_utils import pushd, eval_cmd
+from conda_forge_tick.utils import _get_source_code
 from conda_forge_tick.recipe_parser import CondaMetaYAML
 from conda_forge_tick.migrators.core import MiniMigrator
