+++ conda_forge_tick/migrators/migration_yaml.py
@@ -13,8 +13,8 @@ import networkx as nx
 
 from conda_forge_tick.contexts import FeedstockContext
 from conda_forge_tick.migrators.core import GraphMigrator, MiniMigrator, Migrator
-from conda_forge_tick.utils import pushd
-from conda_forge_tick.utils import eval_cmd, pluck, yaml_safe_load, yaml_safe_dump
+from conda_forge_tick.os_utils import pushd, eval_cmd
+from conda_forge_tick.utils import pluck, yaml_safe_load, yaml_safe_dump
 from conda_forge_tick.make_graph import get_deps_from_outputs_lut
 from conda_forge_tick.feedstock_parser import PIN_SEP_PAT
