+++ conda_forge_tick/update_deps.py
@@ -9,10 +9,8 @@ from typing import Dict
 import requests
 from stdlib_list import stdlib_list
 from grayskull.__main__ import create_python_recipe
-from conda_forge_tick.utils import (
-    _get_source_code,
-    pushd,
-)
+from conda_forge_tick.os_utils import pushd
+from conda_forge_tick.utils import _get_source_code
 from conda_forge_tick.depfinder_api import simple_import_to_pkg_map
 from conda_forge_tick.feedstock_parser import load_feedstock
 from conda_forge_tick.recipe_parser import CondaMetaYAML, CONDA_SELECTOR
