+++ conda_forge_tick/utils.py
@@ -12,15 +12,6 @@ import tempfile
 import io
 import os
 from typing import Any, Tuple, Iterable, Optional, Set
-from concurrent.futures import (
-    ProcessPoolExecutor,
-    ThreadPoolExecutor,
-    Executor,
-)
-import subprocess
-
-from distributed import Lock as DLock
-import multiprocessing
 
 import github3
 import jinja2
@@ -29,26 +20,11 @@ import ruamel.yaml
 import networkx as nx
 
 from . import sensitive_env
+from .lazy_json_backends import LazyJson
 
 if typing.TYPE_CHECKING:
     from mypy_extensions import TypedDict
     from conda_forge_tick.migrators_types import MetaYamlTypedDict
-
-from threading import RLock as TRLock
-
-
-class DummyLock:
-    def __enter__(self, *args, **kwargs):
-        return self
-
-    def __exit__(self, *args, **kwargs):
-        pass
-
-
-TRLOCK = TRLock()
-PRLOCK = DummyLock()
-DLOCK = DummyLock()
-
 
 logger = logging.getLogger("conda_forge_tick.utils")
 
@@ -85,17 +61,6 @@ CB_CONFIG_PINNING = dict(
     cran_mirror="https://cran.r-project.org",
     datetime=datetime,
 )
-
-
-# https://stackoverflow.com/questions/6194499/pushd-through-os-system
-@contextlib.contextmanager
-def pushd(new_dir):
-    previous_dir = os.getcwd()
-    os.chdir(new_dir)
-    try:
-        yield
-    finally:
-        os.chdir(previous_dir)
 
 
 def yaml_safe_load(stream):
@@ -333,29 +298,6 @@ def _parse_meta_yaml_impl(
         raise
 
 
-def eval_cmd(cmd, **kwargs):
-    """run a command capturing stdout
-
-    stderr is printed for debugging
-    any kwargs are added to the env
-    """
-    env = copy.deepcopy(os.environ)
-    timeout = kwargs.pop("timeout", None)
-    env.update(kwargs)
-    c = subprocess.run(
-        cmd,
-        shell=True,
-        stdout=subprocess.PIPE,
-        env=env,
-        timeout=timeout,
-    )
-    if c.returncode != 0:
-        print(c.stdout.decode("utf-8"), flush=True)
-        c.check_returncode()
-
-    return c.stdout.decode("utf-8")
-
-
 class UniversalSet(Set):
     """The universal set, or identity of the set intersection operation."""
 
@@ -426,59 +368,6 @@ def pluck(G: nx.DiGraph, node_id: Any) -> None:
         G.add_edges_from(new_edges)
 
 
-def _init_process(lock):
-    global PRLOCK
-    PRLOCK = lock
-
-
-def _init_dask(lock):
-    global DLOCK
-    DLOCK = lock
-
-
-@contextlib.contextmanager
-def executor(kind: str, max_workers: int, daemon=True) -> typing.Iterator[Executor]:
-    """General purpose utility to get an executor with its as_completed handler
-
-    This allows us to easily use other executors as needed.
-    """
-    global DLOCK
-    global PRLOCK
-
-    if kind == "thread":
-        with ThreadPoolExecutor(max_workers=max_workers) as pool_t:
-            yield pool_t
-    elif kind == "process":
-        m = multiprocessing.Manager()
-        lock = m.RLock()
-        with ProcessPoolExecutor(
-            max_workers=max_workers,
-            initializer=_init_process,
-            initargs=(lock,),
-        ) as pool_p:
-            yield pool_p
-        PRLOCK = DummyLock()
-    elif kind in ["dask", "dask-process", "dask-thread"]:
-        import dask
-        import distributed
-        from distributed.cfexecutor import ClientExecutor
-
-        processes = kind == "dask" or kind == "dask-process"
-
-        with dask.config.set({"distributed.worker.daemon": daemon}):
-            with distributed.LocalCluster(
-                n_workers=max_workers,
-                processes=processes,
-            ) as cluster:
-                with distributed.Client(cluster) as client:
-                    lock = DLock(client=client)
-                    client.run(_init_dask, lock)
-                    yield ClientExecutor(client)
-                DLOCK = DummyLock()
-    else:
-        raise NotImplementedError("That kind is not implemented")
-
-
 def dump_graph_json(gx: nx.DiGraph, filename: str = "graph.json") -> None:
     nld = nx.node_link_data(gx)
     links = nld["links"]
@@ -521,17 +410,17 @@ if typing.TYPE_CHECKING:
 
 
 @typing.overload
-def frozen_to_json_friendly(fz: None, pr: Optional[Any] = None) -> None:
+def frozen_to_json_friendly(fz: None, pr: Optional[LazyJson] = None) -> None:
     pass
 
 
 @typing.overload
-def frozen_to_json_friendly(fz: Any, pr: Optional[Any] = None) -> "JsonFriendly":
+def frozen_to_json_friendly(fz: Any, pr: Optional[LazyJson] = None) -> "JsonFriendly":
     pass
 
 
 @typing.no_type_check
-def frozen_to_json_friendly(fz, pr: Optional[Any] = None):
+def frozen_to_json_friendly(fz, pr: Optional[LazyJson] = None):
     if fz is None:
         return None
     keys = sorted(list(fz.keys()))
