+++ tests/test_security.py
@@ -3,7 +3,7 @@ import subprocess
 
 
 def test_env_is_protected_against_malicious_recipes(tmpdir, caplog, env_setup):
-    from conda_forge_tick.utils import pushd
+    from conda_forge_tick.os_utils import pushd
     import logging
 
     from conda_forge_tick.feedstock_parser import populate_feedstock_attributes
